requirement.txt
streamlit
google-generativeai
yt-dlp
python-dotenv